Fix property optionalism within Input Objects
Within an input object, if the type of a field is nullable, it should
then be optional within typescript as well as allowing `null` as an
explicit value. Conversely, if the type of a field is non-nullable, it
should be both required as a property and not allowed to be `null`.